create workload in morpheus
