+++ examples/morpheusvm/tests/integration/integration_test.go
@@ -4,18 +4,19 @@
 package integration_test
 
 import (
+	"encoding/json"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 
 	"github.com/ava-labs/hypersdk/auth"
 	"github.com/ava-labs/hypersdk/crypto/ed25519"
+	"github.com/ava-labs/hypersdk/examples/morpheusvm/tests/workload"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/vm"
-	"github.com/ava-labs/hypersdk/tests/fixture"
 	"github.com/ava-labs/hypersdk/tests/integration"
 
 	lconsts "github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
-	morpheusGenerator "github.com/ava-labs/hypersdk/examples/morpheusvm/tests/generator"
 	ginkgo "github.com/onsi/ginkgo/v2"
 )
 
@@ -26,10 +27,9 @@ func TestIntegration(t *testing.T) {
 var _ = ginkgo.BeforeSuite(func() {
 	require := require.New(ginkgo.GinkgoT())
 
-	testVM := fixture.NewTestVM(0)
-	keys := testVM.GetKeys()
-	generator := morpheusGenerator.NewTxGenerator(keys[0])
-	genesisBytes, err := testVM.GetGenesisBytes()
+	keys := workload.NewDefaultKeys()
+	genesis := workload.NewGenesis(keys, 100*time.Millisecond)
+	genesisBytes, err := json.Marshal(genesis)
 	require.NoError(err)
 
 	randomEd25519Priv, err := ed25519.GeneratePrivateKey()
@@ -37,6 +37,7 @@ var _ = ginkgo.BeforeSuite(func() {
 
 	randomEd25519AuthFactory := auth.NewED25519Factory(randomEd25519Priv)
 
+	generator := workload.NewTxGenerator(keys[0])
 	// Setup imports the integration test coverage
 	integration.Setup(
 		vm.New,
