+++ examples/morpheusvm/tests/workload/genesis.go
@@ -1,19 +1,31 @@
 // Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package fixture
+package workload
 
 import (
 	"math"
 	"time"
 
 	"github.com/ava-labs/hypersdk/auth"
+	"github.com/ava-labs/hypersdk/codec"
 	"github.com/ava-labs/hypersdk/crypto/ed25519"
 	"github.com/ava-labs/hypersdk/fees"
 	"github.com/ava-labs/hypersdk/genesis"
 )
 
-func newDefaultGenesis(keys []ed25519.PrivateKey, minBlockGap time.Duration) *genesis.DefaultGenesis {
+const (
+	// default initial balance for each address
+	InitialBalance uint64 = 10_000_000_000_000
+)
+
+// hardcoded initial set of ed25519 keys. Each will be initialized with InitialBalance
+var ed25519HexKeys = []string{
+	"323b1d8f4eed5f0da9da93071b034f2dce9d2d22692c172f3cb252a64ddfafd01b057de320297c29ad0c1f589ea216869cf1938d88c9fbd70d6748323dbf2fa7", //nolint:lll
+	"8a7be2e0c9a2d09ac2861c34326d6fe5a461d920ba9c2b345ae28e603d517df148735063f8d5d8ba79ea4668358943e5c80bc09e9b2b9a15b5b15db6c1862e88", //nolint:lll
+}
+
+func NewGenesis(keys []ed25519.PrivateKey, minBlockGap time.Duration) *genesis.DefaultGenesis {
 	// allocate the initial balance to the addresses
 	customAllocs := make([]*genesis.CustomAllocation, 0, len(keys))
 	for _, key := range keys {
@@ -34,4 +46,17 @@ func newDefaultGenesis(keys []ed25519.PrivateKey, minBlockGap time.Duration) *ge
 	genesis.Rules.MinBlockGap = minBlockGap.Milliseconds()
 
 	return genesis
+}
+
+func NewDefaultKeys() []ed25519.PrivateKey {
+	testKeys := make([]ed25519.PrivateKey, len(ed25519HexKeys))
+	for i, keyHex := range ed25519HexKeys {
+		bytes, err := codec.LoadHex(keyHex, ed25519.PrivateKeyLen)
+		if err != nil {
+			panic(err)
+		}
+		testKeys[i] = ed25519.PrivateKey(bytes)
+	}
+
+	return testKeys
 }
