--- tests/fixture/consts.go
@@ -1,15 +1,0 @@
-// Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package fixture
-
-const (
-	// default initial balance for each address
-	InitialBalance uint64 = 10_000_000_000_000
-)
-
-// hardcoded initial set of ed25519 keys. Each will be initialized with InitialBalance
-var ed25519HexKeys = []string{
-	"323b1d8f4eed5f0da9da93071b034f2dce9d2d22692c172f3cb252a64ddfafd01b057de320297c29ad0c1f589ea216869cf1938d88c9fbd70d6748323dbf2fa7", //nolint:lll
-	"8a7be2e0c9a2d09ac2861c34326d6fe5a461d920ba9c2b345ae28e603d517df148735063f8d5d8ba79ea4668358943e5c80bc09e9b2b9a15b5b15db6c1862e88", //nolint:lll
-}
